+++ Hep/ls_listsave/app/src/main/java/com/example/ls_listsave/HashEditText.java
@@ -1,20 +1,32 @@
 package com.example.ls_listsave;
 
 import android.content.Context;
+import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
 import android.text.Editable;
+import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.util.AttributeSet;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.RelativeLayout;
+
+import com.example.ls_listsave.DataBase.LSDBHelper;
+
+import java.util.ArrayList;
+
+import static com.example.ls_listsave.DataBase.LSSQLContract.TagTable.TABLE_NAME;
 
 public class HashEditText extends RelativeLayout {
 
     LayoutInflater inflater = null;
     AutoCompleteTextView editText;
     Button btnClear;
+    ArrayList list = new ArrayList();
     public static Context mContext;
 
     public HashEditText(Context context, AttributeSet attrs) {
@@ -51,6 +63,29 @@ public class HashEditText extends RelativeLayout {
             public void onTextChanged(CharSequence s, int start, int before, int count) {
                 if (s.length() > 0) {
                     btnClear.setVisibility(RelativeLayout.VISIBLE);
+
+                    SQLiteDatabase db = null;
+                    LSDBHelper lsdbHelper = new LSDBHelper(mContext);
+                    db = lsdbHelper.getReadableDatabase();
+
+                    String query = searchSql(s.toString());
+                    Log.d("Search", query);
+                    list.clear();
+                    try {
+                        Cursor cursor = db.rawQuery(query + ";", null);
+                        while (cursor.moveToNext()) {
+                            String result = cursor.getString(2);
+                            list.add(result);
+                            Log.d("Search", result);
+                        }
+                        db.close();
+
+                        AutoCompleteTextView autoCompleteTextView = ((HashEditText) findViewById(R.id.Text_Hash)).editText;
+                        autoCompleteTextView.setAdapter(new ArrayAdapter<String>(mContext, android.R.layout.simple_dropdown_item_1line, list));
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                        db.close();
+                    }
                 } else {
                     btnClear.setVisibility(RelativeLayout.INVISIBLE);
                 }
@@ -63,6 +98,17 @@ public class HashEditText extends RelativeLayout {
                 }
             }
         });
+    }
+
+    public String searchSql(String searchStr) {
+
+        String sql = "Select * FROM " + TABLE_NAME;
+        if (TextUtils.isEmpty(searchStr) == false) {
+            sql += " WHERE ";
+            sql += ChoSearchQuery.makeQuery(searchStr);
+        }
+
+        return sql;
     }
 
     private void clearText() {
