+++ Hep/ls_listsave/app/src/main/res/layout/activity_main.xml
@@ -58,13 +58,6 @@
                 android:onClick="onBtn_ImageAddClicked"
                 android:text="사진추가" />
 
-            <Button
-                android:id="@+id/Search_cho"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="onSearchBtnClicked"
-                android:text="검색" />
-
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
